replace String.indexOf() with contains() Remove unnecessary boxing, unboxing Replace StringBuffer with StringBuilder where applicable Removed an unnecessary ; in SyncHttpClient

+++ src/com/loopj/android/http/BinaryHttpResponseHandler.java
@@ -138,7 +138,7 @@ public class BinaryHttpResponseHandler extends AsyncHttpResponseHandler {
         switch(msg.what) {
             case SUCCESS_MESSAGE:
                 response = (Object[])msg.obj;
-                handleSuccessMessage(((Integer) response[0]).intValue() , (byte[]) response[1]);
+                handleSuccessMessage((Integer) response[0], (byte[]) response[1]);
                 break;
             case FAILURE_MESSAGE:
                 response = (Object[])msg.obj;
